Limiting token database verification when blocks have been pruned
The token database is verified by finding the creation transaction
in the stored block, and then recreating the token creation transaction.
In pruned mode, the block might not be available anymore, so we
limit verification to testing if the block is in the block index if its
data hase been pruned.